% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_fit.R
\name{getCircDiff}
\alias{getCircDiff}
\title{Calculate circular difference}
\usage{
getCircDiff(x, y, period = 1, towardZero = TRUE)
}
\arguments{
\item{x}{Numeric vector or matrix.}

\item{y}{Numeric vector or matrix.}

\item{period}{Period of the periodic variable.}

\item{towardZero}{If \code{TRUE}, returned values will be between \code{-period / 2}
and \code{period / 2}. If \code{FALSE}, returned values will be between 0 and \code{period}.}
}
\value{
Vector or matrix corresponding to \code{x - y}.
}
\description{
\code{getCircDiff} calculates the circular difference.
}
