% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_predict.R
\name{zeitzeigerEnsembleLikelihood}
\alias{zeitzeigerEnsembleLikelihood}
\title{Combine predictions into an ensemble using the log-likelihood.}
\usage{
zeitzeigerEnsembleLikelihood(timeDepLike, timeRange)
}
\arguments{
\item{timeDepLike}{List or 3-D array of time-dependent likelihood from \code{zeitzeigerPredict}.
If a list, then each element (for each member of the ensemble) should be a matrix in which rows
correspond to observations and columns correspond to time-points. If a 3-D array, the three
dimensions should correspond to observations, time-points, and members of the ensemble.}

\item{timeRange}{Vector of time-points at which the likelihood was calculated.}
}
\value{
\item{timeDepLike}{Matrix of likelihood for observations by time-points.}
\item{timePred}{Vector of predicted times. Each predicted time will be an element of timeRange.}
}
\description{
\code{zeitzeigerEnsembleLikelihood} makes predictions by finding the maximum of the sum of the
log-likelihoods.
}

