% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_fit.R
\name{zeitzeigerExtrema}
\alias{zeitzeigerExtrema}
\title{Estimate peaks and troughs.}
\usage{
zeitzeigerExtrema(fitResult, maximum = TRUE, dopar = TRUE)
}
\arguments{
\item{fitResult}{Output of \code{zeitzeigerFit}.}

\item{maximum}{Logical indicating whether to find maximum or minimum.}

\item{dopar}{Logical indicating whether to process features in parallel.
Use \code{\link[doParallel]{registerDoParallel}} to register the parallel backend.}
}
\value{
Matrix with a row for each feature and columns for location and value.
}
\description{
\code{zeitzeigerExtrema} estimates the extremum (peak or trough) for each feature
by using \code{\link[stats]{optimize}} and the periodic spline fit.
}
\seealso{
\code{\link{zeitzeigerFit}}, \code{\link{zeitzeigerSig}}, \code{\link{zeitzeigerSnr}}
}
