% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_fit.R
\name{zeitzeigerLikelihood}
\alias{zeitzeigerLikelihood}
\title{Calculate time-dependent likelihood.}
\usage{
zeitzeigerLikelihood(xTest, xFitMean, xFitVar, beta = NA, timeRange = seq(0,
  1, 0.01), logArg = FALSE)
}
\arguments{
\item{xTest}{Matrix of measurements, with observations in rows and features in columns.}

\item{xFitMean}{List of bigsplines for time-dependent mean, length is number of features.}

\item{xFitVar}{List of bigsplines for time-dependent variance, length is number of features.}

\item{beta}{Vector of coefficients for weighted likelihood. If \code{NA} (default),
then each feature is weighted equally.}

\item{timeRange}{Vector of values of the periodic variable at which to calculate likelihood.}

\item{logArg}{Logical indicating whether to return likilihood (default) or log-likelihood.}
}
\value{
Matrix with observations in rows and times in columns.
}
\description{
Given a matrix of test observations, the estimated time-dependent means
and variances of the features, and a vector of times, \code{zeitzeigerLikelihood}
calculates the likelihood of each time for each test observation. The calculation
assumes that conditioned on the periodic variable, the densities of the features
are normally distributed. The calculation also assumes that the features are
independent.
}

