% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_cv.R
\name{zeitzeigerPredictCv}
\alias{zeitzeigerPredictCv}
\title{Predict corresponding time for observations on cross-validation.}
\usage{
zeitzeigerPredictCv(x, time, foldid, spcResultList, fitMeanArgs = list(rparm =
  NA), constVar = TRUE, fitVarArgs = list(rparm = NA), nSpc = NA,
  betaSv = FALSE, timeRange = seq(0, 1, 0.01))
}
\arguments{
\item{x}{Matrix of measurements, observations in rows and features in columns.}

\item{time}{Vector of values of the periodic variable for observations, where 0
corresponds to the lowest possible value and 1 corresponds to the highest
possible value.}

\item{foldid}{Vector of values indicating which fold each observation is in.}

\item{spcResultList}{Result from \code{zeitzeigerSpcCv}.}

\item{fitMeanArgs}{List of arguments to pass to \code{bigspline} for fitting mean of each SPC.}

\item{constVar}{Logical indicating whether to assume constant variance as a function
of the periodic variable.}

\item{fitVarArgs}{List of arguments to pass to \code{bigspline} for fitting variance of each SPC.
Unused if \code{constVar==TRUE}.}

\item{nSpc}{Vector of the number of SPCs to use for prediction. If \code{NA} (default),
\code{nSpc} will become \code{1:K}, where \code{K} is the number of SPCs in \code{spcResult}.
Each value in \code{nSpc} will correspond to one prediction for each test observation.
A value of 2 means that the prediction will be based on the first 2 SPCs.}

\item{betaSv}{Logical indicating whether to use the singular values of the SPCs
as weights in the likelihood calculation.}

\item{timeRange}{Vector of values of the periodic variable at which to calculate likelihood.
The time with the highest likelihood is used as the initial value for the
MLE optimizer.}
}
\value{
A list of the same structure as \code{zeitzeigerPredict}, combining the results
from each fold of cross-validation.
\item{timeDepLike}{3-D array of likelihood, with dimensions for each observation,
each element of \code{nSpc}, and each element of \code{timeRange}.}
\item{mleFit}{List (for each element in \code{nSpc}) of lists (for each observation)
of \code{mle2} objects.}
\item{timePred}{Matrix of predicted times for observations by values of \code{nSpc}.}
}
\description{
\code{zeitzeigerPredictCv} calls \code{zeitzeigerFit} for each fold of
cross-validation. This function uses \code{doParallel}, so prior to
running this function, use \code{registerDoParallel} to set the number of cores.
}

