% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_cut.R
\name{zeitzeigerProp}
\alias{zeitzeigerProp}
\title{Calculate the rhythmic properties of each feature}
\usage{
zeitzeigerProp(fitResult, dopar = TRUE)
}
\arguments{
\item{fitResult}{Output of \code{\link[=zeitzeigerFit]{zeitzeigerFit()}}.}

\item{dopar}{Logical indicating whether to process features in parallel.
Use \code{\link[doParallel:registerDoParallel]{doParallel::registerDoParallel()}} to register the parallel backend.}
}
\value{
\code{data.frame} with a row for each feature.
}
\description{
DEPRECATED: We recommend instead using
\href{https://github.com/hugheylab/limorhyde}{limorhyde}, which has support for
cosinor and periodic splines, in combination with methods such as
\href{https://doi.org/doi:10.18129/B9.bioc.limma}{limma}.
}
\details{
Calculate the rhythmic properties of each feature's spline fit: location and
value of peak, location and value of trough, amplitude measured peak to
trough, and signal-to-noise ratio (amplitude divided by the square root of
the mean of the squared residuals).
}
\seealso{
\code{\link[=zeitzeigerFit]{zeitzeigerFit()}}
}
