% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_fit.R
\name{zeitzeigerSpc}
\alias{zeitzeigerSpc}
\title{Calculate sparse principal components of time-dependent variation}
\usage{
zeitzeigerSpc(
  xFitMean,
  xFitResid,
  nTime = 10,
  useSpc = TRUE,
  sumabsv = 1,
  orth = TRUE,
  ...
)
}
\arguments{
\item{xFitMean}{List of bigsplines, length is number of features.}

\item{xFitResid}{Matrix of residuals, dimensions are observations by
features.}

\item{nTime}{Number of time-points by which to discretize the time-dependent
behavior of each feature. Corresponds to the number of rows in the matrix
for which the SPCs will be calculated.}

\item{useSpc}{Logical indicating whether to use \code{\link[PMA:SPC]{PMA::SPC()}} (default) or
\code{\link[base:svd]{base::svd()}}.}

\item{sumabsv}{L1-constraint on the SPCs, passed to \code{\link[PMA:SPC]{PMA::SPC()}}.}

\item{orth}{Logical indicating whether to require left singular vectors
be orthogonal to each other, passed to \code{\link[PMA:SPC]{PMA::SPC()}}.}

\item{...}{Other arguments passed to \code{\link[PMA:SPC]{PMA::SPC()}}.}
}
\value{
Output of \code{\link[PMA:SPC]{PMA::SPC()}}, unless \code{useSpc} is \code{FALSE}, then output of
\code{\link[base:svd]{base::svd()}}.
}
\description{
Calculate the SPCs given the time-dependent means and the residuals from
\code{\link[=zeitzeigerFit]{zeitzeigerFit()}}.
}
\seealso{
\code{\link[=zeitzeigerFit]{zeitzeigerFit()}}, \code{\link[=zeitzeigerPredict]{zeitzeigerPredict()}}
}
