% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_cv.R
\name{zeitzeigerSpcCv}
\alias{zeitzeigerSpcCv}
\title{Calculate sparse principal components of time-dependent variation
on cross-validation.}
\usage{
zeitzeigerSpcCv(fitResultList, nTime = 10, useSpc = TRUE, sumabsv = 1,
  orth = TRUE)
}
\arguments{
\item{fitResultList}{Result from \code{zeitzeigerFitCv}.}

\item{nTime}{Number of time-points by which to discretize the time-dependent
behavior of each feature. Corresponds to the number of rows in the matrix for
which the SPCs will be calculated.}

\item{sumabsv}{L1-constraint on the SPCs, passed to \code{SPC}.}

\item{orth}{Logical indicating whether to require left singular vectors
be orthogonal to each other, passed to \code{SPC}.}

\item{useSPC}{Logical indicating whether to use \code{SPC} (default) or \code{svd}.}
}
\value{
A list consisting of the result from \code{zeitzeigerSpc} for each fold.
}
\description{
\code{zeitzeigerSpcCv} calls \code{zeitzeigerFit} for each fold of
cross-validation. This function uses \code{doParallel}, so prior to
running this function, use \code{registerDoParallel} to set the number of cores.
}

