% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_predict.R
\name{zeitzeiger}
\alias{zeitzeiger}
\title{Train and test a ZeitZeiger predictor.}
\usage{
zeitzeiger(xTrain, timeTrain, xTest, fitMeanArgs = list(rparm = NA,
  nknots = 3), constVar = TRUE, fitVarArgs = list(rparm = NA),
  nTime = 10, useSpc = TRUE, sumabsv = 2, orth = TRUE, nSpc = 2,
  betaSv = FALSE, timeRange = seq(0, 1, 0.01))
}
\arguments{
\item{xTrain}{Matrix of measurements for training data, observations in rows
and features in columns.}

\item{timeTrain}{Vector of values of the periodic variable for training observations,
where 0 corresponds to the lowest possible value and 1 corresponds to the highest
possible value.}

\item{xTest}{Matrix of measurements for test data, observations in rows
and features in columns.}

\item{fitMeanArgs}{List of arguments to pass to \code{bigspline} for fitting mean of each SPC.}

\item{constVar}{Logical indicating whether to assume constant variance as a function
of the periodic variable.}

\item{fitVarArgs}{List of arguments to pass to \code{bigspline} for fitting variance of each SPC.
Unused if \code{constVar==TRUE}.}

\item{nTime}{Number of time-points by which to discretize the time-dependent
behavior of each feature. Corresponds to the number of rows in the matrix for
which the SPCs will be calculated.}

\item{useSpc}{Logical indicating whether to use \code{SPC} (default) or \code{svd}.}

\item{sumabsv}{L1-constraint on the SPCs, passed to \code{SPC}.}

\item{orth}{Logical indicating whether to require left singular vectors
be orthogonal to each other, passed to \code{SPC}.}

\item{nSpc}{Vector of the number of SPCs to use for prediction. If \code{NA} (default),
\code{nSpc} will become \code{1:K}, where \code{K} is the number of SPCs in \code{spcResult}.
Each value in \code{nSpc} will correspond to one prediction for each test observation.
A value of 2 means that the prediction will be based on the first 2 SPCs.}

\item{betaSv}{Logical indicating whether to use the singular values of the SPCs
as weights in the likelihood calculation.}

\item{timeRange}{Vector of values of the periodic variable at which to calculate likelihood.
The time with the highest likelihood is used as the initial value for the
MLE optimizer.}
}
\value{
\item{fitResult}{Result from \code{zeitzeigerFit}}
\item{spcResult}{Result from \code{zeitzeigerSpc}}
\item{predResult}{Result from \code{zeitzeigerPredict}}
}
\description{
\code{zeitzeiger} sequentially calls \code{zeitzeigerFit}, \code{zeitzeigerSpc},
and \code{zeitzeigerPredict}.
}
\seealso{
\code{\link{zeitzeigerFit}}, \code{\link{zeitzeigerSpc}}, \code{\link{zeitzeigerPredict}}
}
