% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_cut.R
\name{zeitzeigerExtrema}
\alias{zeitzeigerExtrema}
\title{Estimate peaks and troughs}
\usage{
zeitzeigerExtrema(fitResult, maximum = TRUE, dopar = TRUE)
}
\arguments{
\item{fitResult}{Output of \code{zeitzeigerFit}.}

\item{maximum}{Logical indicating whether to find maximum or minimum.}

\item{dopar}{Logical indicating whether to process features in parallel.
Use \code{\link[doParallel]{registerDoParallel}} to register the parallel backend.}
}
\value{
Matrix with a row for each feature and columns for location and value.
}
\description{
DEPRECATED: We recommend instead using
\href{https://github.com/hugheylab/limorhyde}{limorhyde}, which has support for
cosinor and periodic splines, in combination with methods such as
\href{https://doi.org/doi:10.18129/B9.bioc.limma}{limma}.
}
\details{
\code{zeitzeigerExtrema} estimates the extremum (peak or trough) for each feature
by using \code{\link[stats]{optimize}} and the periodic spline fit. Directly calling
this function is deprecated. Instead, use \code{\link{zeitzeigerProp}}.
}
\seealso{
\code{\link{zeitzeigerFit}}, \code{\link{zeitzeigerProp}}
}
