% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_fit.R
\name{zeitzeigerFit}
\alias{zeitzeigerFit}
\title{Estimate time-dependent mean.}
\usage{
zeitzeigerFit(x, time, fitMeanArgs = list(rparm = NA))
}
\arguments{
\item{x}{Matrix of measurements, with observations in rows and features in columns.
Missing values are allowed.}

\item{time}{Vector of values of the periodic variable for the observations, where 0
corresponds to the lowest possible value and 1 corresponds to the highest possible value.}

\item{fitMeanArgs}{List of arguments to pass to \code{bigspline}.}
}
\value{
\item{xFitMean}{List of results from \code{bigspline}. Length is number of columns in \code{x}.}
\item{xFitResid}{Matrix of residuals, same dimensions as \code{x}.}
}
\description{
\code{zeitzeigerFit} estimates the mean of each feature as a function of the
periodic variable, using a periodic smoothing spline.
}

