% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_cv.R
\name{zeitzeigerFitCv}
\alias{zeitzeigerFitCv}
\title{Estimate time-dependent mean on cross-validation.}
\usage{
zeitzeigerFitCv(x, time, foldid, fitMeanArgs = list(rparm = NA),
  dopar = TRUE)
}
\arguments{
\item{x}{Matrix of measurements, with observations in rows and features in columns.}

\item{time}{Vector of values of the periodic variable for the observations, where 0
corresponds to the lowest possible value and 1 corresponds to the highest possible value.}

\item{foldid}{Vector of values indicating which fold each observation is in.}

\item{fitMeanArgs}{List of arguments to pass to \code{bigspline}.}

\item{dopar}{Logical indicating whether to process the folds in parallel. Before calling
this function, use \code{doParallel::registerDoParallel} to register the parallel backend.}
}
\value{
A list consisting of the result from \code{zeitzeigerFit} for each fold.
}
\description{
\code{zeitzeigerFitCv} calls \code{zeitzeigerFit} for each fold of cross-validation.
By default, if a parallel backend is registered, this function processes the folds in parallel.
}
\seealso{
\code{\link{zeitzeigerFit}}, \code{\link{zeitzeigerSpcCv}}, \code{\link{zeitzeigerPredictCv}}
}

