% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_cv.R
\name{zeitzeigerFitCv}
\alias{zeitzeigerFitCv}
\title{Fit a periodic spline for each feature on cross-validation}
\usage{
zeitzeigerFitCv(x, time, foldid, nKnots = 3)
}
\arguments{
\item{x}{Matrix of measurements, with observations in rows and features in
columns.}

\item{time}{Vector of values of the periodic variable for the observations,
where 0 corresponds to the lowest possible value and 1 corresponds to the
highest possible value.}

\item{foldid}{Vector of values indicating the fold to which each observation
belongs.}

\item{nKnots}{Number of internal knots to use for the periodic smoothing
spline.}
}
\value{
A list consisting of the result from \code{\link[=zeitzeigerFit]{zeitzeigerFit()}} for each fold.
}
\description{
Fit a periodic spline for each feature for each fold of cross-validation.
}
\seealso{
\code{\link[=zeitzeigerFit]{zeitzeigerFit()}}, \code{\link[=zeitzeigerSpcCv]{zeitzeigerSpcCv()}}, \code{\link[=zeitzeigerPredictCv]{zeitzeigerPredictCv()}}
}
