% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_fit.R
\name{zeitzeigerProp}
\alias{zeitzeigerProp}
\title{Calculate the rhythmic properties of each feature.}
\usage{
zeitzeigerProp(fitResult, dopar = TRUE)
}
\arguments{
\item{fitResult}{Output of \code{zeitzeigerFit}.}

\item{dopar}{Logical indicating whether to process features in parallel.
Use \code{\link[doParallel]{registerDoParallel}} to register the parallel backend.}
}
\value{
Data frame with a row for each feature.
}
\description{
\code{zeitzeigerProp} calculates the rhythmic properties of each feature's
spline fit: location and value of peak, location and value of trough,
amplitude measured peak to trough, and signal-to-noise ratio (amplitude
divided by the square root of the mean of the squared residuals).
}
\seealso{
\code{\link{zeitzeigerFit}}
}
