% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitzeiger_cut.R
\name{zeitzeigerSnr}
\alias{zeitzeigerSnr}
\title{Calculate the signal-to-noise of the periodic spline fits}
\usage{
zeitzeigerSnr(fitResult, dopar = TRUE)
}
\arguments{
\item{fitResult}{Output of \code{\link[=zeitzeigerFit]{zeitzeigerFit()}}.}

\item{dopar}{Logical indicating whether to process features in parallel.
Use \code{\link[doParallel:registerDoParallel]{doParallel::registerDoParallel()}} to register the parallel backend.}
}
\value{
Vector of signal-to-noise values.
}
\description{
DEPRECATED: We recommend instead using
\href{https://github.com/hugheylab/limorhyde}{\code{limorhyde}}, which has support for
cosinor and periodic splines, in combination with methods such as
\href{https://doi.org/doi:10.18129/B9.bioc.limma}{\code{limma}}.
}
\details{
Calculate the signal-to-noise ratio of the spline fit for each feature. The
SNR is calculated as the peak-to-trough difference, divided by the square
root of the mean of the squared residuals.
}
\seealso{
\code{\link[=zeitzeigerFit]{zeitzeigerFit()}}, \code{\link[=zeitzeigerProp]{zeitzeigerProp()}}
}
